Added -loader suffix to both examples
With webpack the `-loader` bit in the loader name is optional and even configurable, but nobody needs to know that to get started.

I think mixing usage of the optional-ness of it will confuse people, so this example's string should either both have -loader or neither. This MR adds -loader.